Reimplement systemd handling.
Following suggested changes from Lennart.  This patch set will mount
a tmpfs on /sys/fs/systemd inside of the container if it is not already
mounted.  If it is already mounted it will mount the scope directory from
the host for the running container into the container in the same location

grep systemd /proc/self/cgroup
1:name=systemd:/system.slice/docker-325a07e48b91822d5ad837ef4117355dcd2b9389daa148585f5083de7580570a.scope

The code would take this to generate the path

/sys/fs/cgroup/systemd/system.slice/docker-325a07e48b91822d5ad837ef4117355dcd2b9389daa148585f5083de7580570a.scope

And then mount this on
$ROOTFS/sys/fs/cgroup/systemd/system.slice/docker-325a07e48b91822d5ad837ef4117355dcd2b9389daa148585f5083de7580570a.scope

Finally it will change the labels on this directory to match the mount label of
the container.

This will allow systemd inside of the container to write to only the appropriate
directory on the host.